requirements.txt
torch
torchvision
Pillow
tqdm
fastapi
uvicorn
python-multipart
